requirements.txt.txt
django
django-allauth
django-import-export
django-widget-tweaks
django-simple-captcha